Update writeData() return type to float

+++ PressureTransducer.h
@@ -13,7 +13,7 @@ public:
                        float rangeMin, float rangeMax, float calibrationOffset = 0.0f);
 
     float readData() override;
-    int writeData(float data) override;
+    float writeData(float data) override;
 };
 
 #endif // PRESSURE_TRANSDUCER_H
